+++ lib/actions/product.actions.ts
@@ -1,10 +1,11 @@
 "use server";
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { CompanyName, CompanyProductCount, PhoneNumber, Sort } from "@/types";
+import { CompanyName, PhoneNumber, Result, Sort } from "@/types";
 import Product from "../models/products.model";
 import { connectToDatabase } from "../mongoose";
 import { revalidateTag, unstable_cache } from "next/cache";
 import { tags } from "@/constants";
+import Offer from "../models/offer.model";
 
 interface FetchOptions {
   page?: number;
@@ -230,18 +231,26 @@ export const insertManyProducts = async (
   }
 };
 
-export const addProduct = async (product: Omit<PhoneNumber, "_id">) => {
+export const addProduct = async (
+  product: Omit<PhoneNumber, "_id">,
+): Promise<Result> => {
   try {
     await connectToDatabase();
     await Product.create(product);
     revalidateTag(tags.products);
+    return { success: true, message: "Product added successfully." };
   } catch (error: any) {
-    throw new Error(`Error adding product: ${error.message}`);
+    console.error("Error adding product:", error);
+    if (error.message.includes("duplicate")) {
+      return { success: false, message: "هذا الرقم موجود بالفعل" };
+    } else {
+      return { success: false, message: "حدث خطأ في إضافة المنتج" };
+    }
   }
 };
 
 // Update a product by ID
-export const updateProductById = async (data: PhoneNumber) => {
+export const updateProductById = async (data: PhoneNumber): Promise<Result> => {
   const { _id, ...updateData } = data;
   const updatedData = Object.fromEntries(
     Object.entries(updateData).filter(
@@ -251,31 +260,48 @@ export const updateProductById = async (data: PhoneNumber) => {
   );
   try {
     await connectToDatabase();
+    if (updateData.active === false) {
+      const isInOffers = await Offer.findOne({
+        phoneNumber: updateData.phoneNumber,
+        active: true,
+      });
+      if (isInOffers) {
+        return {
+          success: false,
+          message:
+            " لا يمكنك عدم تفعيل هذا الرقم لانه دتخل عرض فعال قم بتعطيله اولا",
+        };
+      }
+    }
     await Product.findByIdAndUpdate(_id, updatedData);
     revalidateTag(tags.products);
+    return { success: true, message: "Product updated successfully." };
   } catch (error: any) {
-    throw new Error(`Error updating product with ID ${_id}: ${error.message}`);
-  }
-};
-
-export const switchProductActive = async (data: PhoneNumber) => {
-  const { _id, ...updateData } = data;
-  try {
-    await connectToDatabase();
-    await Product.findOneAndUpdate({ _id }, { active: !updateData.active });
-    revalidateTag(tags.products);
-  } catch (error: any) {
-    throw new Error(`Error updating product with ID ${_id}: ${error.message}`);
+    console.error("Error updating product:", error);
+    if (error.message.includes("duplicate")) {
+      return { success: false, message: "هذا الرقم موجود بالفعل" };
+    } else {
+      return { success: false, message: "حدث خطأ في تحديث المنتج" };
+    }
   }
 };
 
 // Delete a product by ID
-export const deleteProductById = async (item: PhoneNumber) => {
+export const deleteProductById = async (item: PhoneNumber): Promise<Result> => {
   const id = item._id;
   try {
     await connectToDatabase();
+    const isInOffers = await Offer.findOne({ phoneNumber: item.phoneNumber });
+    if (isInOffers) {
+      return {
+        success: false,
+        message: " لا يمكنك حذف هذا الرقم لانه دتخل عرض فعال قم بتعطيله اولا",
+      };
+    }
+
     await Product.findByIdAndDelete(id);
     revalidateTag(tags.products);
+    return { success: true, message: "Product deleted successfully." };
   } catch (error: any) {
     throw new Error(`Error deleting product with ID ${id}: ${error.message}`);
   }
@@ -294,7 +320,9 @@ export const fetchProductById = async (id: string) => {
 export const fetchProductByPhoneNumber = async (phoneNumber: string) => {
   try {
     await connectToDatabase();
-    return await Product.findOne({ phoneNumber });
+    const data = await Product.findOne({ phoneNumber });
+    const product: PhoneNumber = JSON.parse(JSON.stringify(data));
+    return product;
   } catch (error: any) {
     throw new Error(
       `Error fetching product with phoneNumber ${phoneNumber}: ${error.message}`,
@@ -302,25 +330,21 @@ export const fetchProductByPhoneNumber = async (phoneNumber: string) => {
   }
 };
 
-export const countProductsByCompanies = unstable_cache(
-  async (companyNames: string[]): Promise<CompanyProductCount[]> => {
-    try {
-      await connectToDatabase();
-      const productCounts = await Product.aggregate([
-        { $match: { company: { $in: companyNames }, active: true } },
-        { $group: { _id: "$company", count: { $sum: 1 } } },
-        { $project: { company: "$_id", count: 1, _id: 0 } },
-      ]);
-      return productCounts;
-    } catch (error: any) {
-      throw new Error(
-        `Error counting products for companies ${companyNames.join(", ")}: ${error.message}`,
-      );
-    }
-  },
-  [tags.productCounts, tags.products],
-  {
-    tags: [tags.productCounts, tags.products],
-    revalidate: 60 * 60 * 24 * 7, // 7 days
-  },
-);
+export const fetchProductsByPhoneNumber = async (query: string) => {
+  const searchCriteria = query
+    ? {
+        $or: [{ phoneNumber: { $regex: `^${query}`, $options: "i" } }],
+      }
+    : {};
+
+  try {
+    await connectToDatabase();
+    const data = await Product.findOne(searchCriteria);
+    const product: PhoneNumber = JSON.parse(JSON.stringify(data));
+    return product;
+  } catch (error: any) {
+    throw new Error(
+      `Error fetching product with phoneNumber ${query}: ${error.message}`,
+    );
+  }
+};
